learn c pointer type

+++ me.h
@@ -1,1 +1,1 @@
-int get_age();
+float get_age();
